# api/v1/endpoints/history.py
# -*- coding: utf-8 -*-
"""
===================================
历史记录接口
===================================

职责：
1. 提供 GET /api/v1/history 历史列表查询接口
2. 提供 GET /api/v1/history/{query_id} 历史详情查询接口
"""

import logging
from typing import Any, Dict, List, Optional

from fastapi import APIRouter, HTTPException, Query, Depends, Body

from api.deps import get_database_manager, optional_portal_user_id
from api.v1.schemas.history import (
    HistoryListResponse,
    HistoryItem,
    DeleteHistoryRequest,
    DeleteHistoryResponse,
    LatestAnalysisSummariesResponse,
    LatestAnalysisSummaryItem,
    NewsIntelItem,
    NewsIntelResponse,
    AnalysisReport,
    ReportMeta,
    ReportSummary,
    ReportStrategy,
    ReportDetails,
    MarkdownReportResponse,
)
from api.v1.schemas.common import ErrorResponse
from src.storage import DatabaseManager
from src.report_language import (
    get_sentiment_label,
    get_localized_stock_name,
    localize_operation_advice,
    localize_trend_prediction,
    normalize_report_language,
)
from src.services.history_service import HistoryService, MarkdownReportGenerationError
from src.utils.data_processing import (
    normalize_model_used,
    extract_fundamental_detail_fields,
    extract_board_detail_fields,
)

logger = logging.getLogger(__name__)

router = APIRouter()


def build_analysis_report_from_detail_dict(
    result: Dict[str, Any],
    db_manager: DatabaseManager,
) -> AnalysisReport:
    """
    将 HistoryService.resolve_and_get_detail / _record_to_detail_dict 的结果转为 API 层 AnalysisReport。
    与 GET /api/v1/history/{id} 响应字段一致，供公开示例接口复用。
    """
    current_price = None
    change_pct = None
    context_snapshot = result.get("context_snapshot")
    if context_snapshot and isinstance(context_snapshot, dict):
        enhanced_context = context_snapshot.get("enhanced_context") or {}
        realtime = enhanced_context.get("realtime") or {}
        current_price = realtime.get("price")
        change_pct = realtime.get("change_pct") or realtime.get("change_60d")

        if current_price is None:
            realtime_quote_raw = context_snapshot.get("realtime_quote_raw") or {}
            current_price = realtime_quote_raw.get("price")
            change_pct = change_pct or realtime_quote_raw.get("change_pct") or realtime_quote_raw.get("pct_chg")

    raw_result = result.get("raw_result")
    if not isinstance(raw_result, dict):
        raw_result = {}
    report_language = normalize_report_language(
        result.get("report_language")
        or raw_result.get("report_language")
        or (
            context_snapshot.get("report_language")
            if isinstance(context_snapshot, dict)
            else None
        )
    )
    stock_name = get_localized_stock_name(
        result.get("stock_name"),
        result.get("stock_code", ""),
        report_language,
    )

    meta = ReportMeta(
        id=result.get("id"),
        query_id=result.get("query_id", ""),
        stock_code=result.get("stock_code", ""),
        stock_name=stock_name,
        report_type=result.get("report_type"),
        report_language=report_language,
        created_at=result.get("created_at"),
        current_price=current_price,
        change_pct=change_pct,
        model_used=normalize_model_used(result.get("model_used")),
    )

    summary = ReportSummary(
        analysis_summary=result.get("analysis_summary"),
        operation_advice=localize_operation_advice(
            result.get("operation_advice"),
            report_language,
        ),
        trend_prediction=localize_trend_prediction(
            result.get("trend_prediction"),
            report_language,
        ),
        sentiment_score=result.get("sentiment_score"),
        sentiment_label=(
            get_sentiment_label(result.get("sentiment_score"), report_language)
            if result.get("sentiment_score") is not None
            else result.get("sentiment_label")
        ),
    )

    strategy = ReportStrategy(
        ideal_buy=result.get("ideal_buy"),
        secondary_buy=result.get("secondary_buy"),
        stop_loss=result.get("stop_loss"),
        take_profit=result.get("take_profit"),
    )

    fallback_fundamental = db_manager.get_latest_fundamental_snapshot(
        query_id=result.get("query_id", ""),
        code=result.get("stock_code", ""),
    )
    extracted_fundamental = extract_fundamental_detail_fields(
        context_snapshot=result.get("context_snapshot"),
        fallback_fundamental_payload=fallback_fundamental,
    )
    extracted_boards = extract_board_detail_fields(
        context_snapshot=result.get("context_snapshot"),
        fallback_fundamental_payload=fallback_fundamental,
    )

    details = ReportDetails(
        news_content=result.get("news_content"),
        raw_result=result.get("raw_result"),
        context_snapshot=result.get("context_snapshot"),
        financial_report=extracted_fundamental.get("financial_report"),
        dividend_metrics=extracted_fundamental.get("dividend_metrics"),
        belong_boards=extracted_boards.get("belong_boards"),
        sector_rankings=extracted_boards.get("sector_rankings"),
    )

    return AnalysisReport(
        meta=meta,
        summary=summary,
        strategy=strategy,
        details=details,
    )


@router.get(
    "",
    response_model=HistoryListResponse,
    responses={
        200: {"description": "历史记录列表"},
        500: {"description": "服务器错误", "model": ErrorResponse},
    },
    summary="获取历史分析列表",
    description=(
        "分页获取历史分析记录摘要，支持按股票代码和日期范围筛选。"
        "``mine=true`` 时仅返回当前门户 Cookie 对应用户提交的分析（需已登录）；"
        "``q`` 为代码或名称模糊匹配。"
    ),
)
def get_history_list(
    stock_code: Optional[str] = Query(None, description="股票代码筛选"),
    start_date: Optional[str] = Query(None, description="开始日期 (YYYY-MM-DD)"),
    end_date: Optional[str] = Query(None, description="结束日期 (YYYY-MM-DD)"),
    page: int = Query(1, ge=1, description="页码（从 1 开始）"),
    limit: int = Query(20, ge=1, le=100, description="每页数量"),
    mine: bool = Query(False, description="仅当前门户用户提交的记录"),
    q: Optional[str] = Query(None, description="股票代码或名称关键词（模糊）"),
    portal_uid: Optional[int] = Depends(optional_portal_user_id),
    db_manager: DatabaseManager = Depends(get_database_manager)
) -> HistoryListResponse:
    """
    获取历史分析列表
    
    分页获取历史分析记录摘要，支持按股票代码和日期范围筛选
    
    Args:
        stock_code: 股票代码筛选
        start_date: 开始日期
        end_date: 结束日期
        page: 页码
        limit: 每页数量
        db_manager: 数据库管理器依赖
        
    Returns:
        HistoryListResponse: 历史记录列表
    """
    try:
        service = HistoryService(db_manager)
        
        # 使用 def 而非 async def，FastAPI 自动在线程池中执行
        result = service.get_history_list(
            stock_code=stock_code,
            start_date=start_date,
            end_date=end_date,
            page=page,
            limit=limit,
            mine_only=mine,
            portal_user_id=portal_uid,
            search_q=q,
        )
        
        # 转换为响应模型
        items = [
            HistoryItem(
                id=item.get("id"),
                query_id=item.get("query_id", ""),
                stock_code=item.get("stock_code", ""),
                stock_name=item.get("stock_name"),
                report_type=item.get("report_type"),
                sentiment_score=item.get("sentiment_score"),
                operation_advice=item.get("operation_advice"),
                concept_tags=item.get("concept_tags") or [],
                created_at=item.get("created_at")
            )
            for item in result.get("items", [])
        ]
        
        return HistoryListResponse(
            total=result.get("total", 0),
            page=page,
            limit=limit,
            items=items
        )
        
    except Exception as e:
        logger.error(f"查询历史列表失败: {e}", exc_info=True)
        raise HTTPException(
            status_code=500,
            detail={
                "error": "internal_error",
                "message": f"查询历史列表失败: {str(e)}"
            }
        )


@router.delete(
    "",
    response_model=DeleteHistoryResponse,
    responses={
        200: {"description": "删除成功"},
        400: {"description": "请求参数错误", "model": ErrorResponse},
        500: {"description": "服务器错误", "model": ErrorResponse},
    },
    summary="删除历史分析记录",
    description="按历史记录主键 ID 批量删除分析历史"
)
def delete_history_records(
    request: DeleteHistoryRequest = Body(...),
    db_manager: DatabaseManager = Depends(get_database_manager)
) -> DeleteHistoryResponse:
    """
    按主键 ID 批量删除历史分析记录。
    """
    record_ids = sorted({record_id for record_id in request.record_ids if record_id is not None})
    if not record_ids:
        raise HTTPException(
            status_code=400,
            detail={
                "error": "invalid_request",
                "message": "record_ids 不能为空"
            }
        )

    try:
        service = HistoryService(db_manager)
        deleted = service.delete_history_records(record_ids)
        return DeleteHistoryResponse(deleted=deleted)
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"删除历史记录失败: {e}", exc_info=True)
        raise HTTPException(
            status_code=500,
            detail={
                "error": "internal_error",
                "message": f"删除历史记录失败: {str(e)}"
            }
        )


@router.get(
    "/latest-summaries",
    response_model=LatestAnalysisSummariesResponse,
    responses={
        200: {"description": "按代码返回最近一条分析摘要"},
        500: {"description": "服务器错误", "model": ErrorResponse},
    },
    summary="批量获取最近分析摘要",
    description=(
        "按股票代码批量查询其在分析历史中「创建时间最新」的一条记录摘要："
        "评分、操作建议与概念板块名。用于自选等列表展示。"
        "传入 `codes` 查询参数（可重复）或逗号分隔的单个 `codes`。"
    ),
)
def get_latest_analysis_summaries(
    codes: List[str] = Query(
        default_factory=list,
        description="股票代码；可重复 `?codes=A&codes=B`，或与逗号合用",
        alias="codes",
    ),
    db_manager: DatabaseManager = Depends(get_database_manager),
) -> LatestAnalysisSummariesResponse:
    flat: List[str] = []
    for part in codes or []:
        for seg in str(part).split(","):
            s = seg.strip()
            if s:
                flat.append(s)

    try:
        service = HistoryService(db_manager)
        raw_items = service.get_latest_summaries_for_codes(flat, max_codes=200)
        items = {
            k: LatestAnalysisSummaryItem(**v)
            for k, v in raw_items.items()
        }
        return LatestAnalysisSummariesResponse(items=items)
    except Exception as e:
        logger.error(f"批量查询最近分析摘要失败: {e}", exc_info=True)
        raise HTTPException(
            status_code=500,
            detail={
                "error": "internal_error",
                "message": f"批量查询最近分析摘要失败: {str(e)}",
            },
        )


@router.get(
    "/{record_id}",
    response_model=AnalysisReport,
    responses={
        200: {"description": "报告详情"},
        404: {"description": "报告不存在", "model": ErrorResponse},
        500: {"description": "服务器错误", "model": ErrorResponse},
    },
    summary="获取历史报告详情",
    description="根据分析历史记录 ID 或 query_id 获取完整的历史分析报告"
)
def get_history_detail(
    record_id: str,
    db_manager: DatabaseManager = Depends(get_database_manager)
) -> AnalysisReport:
    """
    获取历史报告详情
    
    根据分析历史记录主键 ID 或 query_id 获取完整的历史分析报告。
    优先尝试按主键 ID（整数）查询，若参数不是合法整数则按 query_id 查询。
    
    Args:
        record_id: 分析历史记录主键 ID（整数）或 query_id（字符串）
        db_manager: 数据库管理器依赖
        
    Returns:
        AnalysisReport: 完整分析报告
        
    Raises:
        HTTPException: 404 - 报告不存在
    """
    try:
        service = HistoryService(db_manager)
        
        # Try integer ID first, fall back to query_id string lookup
        result = service.resolve_and_get_detail(record_id)

        if result is None:
            raise HTTPException(
                status_code=404,
                detail={
                    "error": "not_found",
                    "message": f"未找到 id/query_id={record_id} 的分析记录"
                }
            )

        return build_analysis_report_from_detail_dict(result, db_manager)
        
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"查询历史详情失败: {e}", exc_info=True)
        raise HTTPException(
            status_code=500,
            detail={
                "error": "internal_error",
                "message": f"查询历史详情失败: {str(e)}"
            }
        )


@router.get(
    "/{record_id}/news",
    response_model=NewsIntelResponse,
    responses={
        200: {"description": "新闻情报列表"},
        500: {"description": "服务器错误", "model": ErrorResponse},
    },
    summary="获取历史报告关联新闻",
    description="根据分析历史记录 ID 获取关联的新闻情报列表（为空也返回 200）"
)
def get_history_news(
    record_id: str,
    limit: int = Query(20, ge=1, le=100, description="返回数量限制"),
    db_manager: DatabaseManager = Depends(get_database_manager)
) -> NewsIntelResponse:
    """
    获取历史报告关联新闻

    根据分析历史记录 ID 或 query_id 获取关联的新闻情报列表。
    在内部完成 record_id → query_id 的解析。

    Args:
        record_id: 分析历史记录主键 ID（整数）或 query_id（字符串）
        limit: 返回数量限制
        db_manager: 数据库管理器依赖

    Returns:
        NewsIntelResponse: 新闻情报列表
    """
    try:
        service = HistoryService(db_manager)
        items = service.resolve_and_get_news(record_id=record_id, limit=limit)

        response_items = [
            NewsIntelItem(
                title=item.get("title", ""),
                snippet=item.get("snippet"),
                url=item.get("url", "")
            )
            for item in items
        ]

        return NewsIntelResponse(
            total=len(response_items),
            items=response_items
        )

    except Exception as e:
        logger.error(f"查询新闻情报失败: {e}", exc_info=True)
        raise HTTPException(
            status_code=500,
            detail={
                "error": "internal_error",
                "message": f"查询新闻情报失败: {str(e)}"
            }
        )


@router.get(
    "/{record_id}/markdown",
    response_model=MarkdownReportResponse,
    responses={
        200: {"description": "Markdown 格式报告"},
        404: {"description": "报告不存在", "model": ErrorResponse},
        500: {"description": "服务器错误", "model": ErrorResponse},
    },
    summary="获取历史报告 Markdown 格式",
    description="根据分析历史记录 ID 获取 Markdown 格式的完整分析报告"
)
def get_history_markdown(
    record_id: str,
    db_manager: DatabaseManager = Depends(get_database_manager)
) -> MarkdownReportResponse:
    """
    获取历史报告的 Markdown 格式内容

    根据分析历史记录 ID 或 query_id 生成与推送通知格式一致的 Markdown 报告。

    Args:
        record_id: 分析历史记录主键 ID（整数）或 query_id（字符串）
        db_manager: 数据库管理器依赖

    Returns:
        MarkdownReportResponse: Markdown 格式的完整报告

    Raises:
        HTTPException: 404 - 报告不存在
        HTTPException: 500 - 报告生成失败（服务器内部错误）
    """
    service = HistoryService(db_manager)

    try:
        markdown_content = service.get_markdown_report(record_id)
    except MarkdownReportGenerationError as e:
        logger.error(f"Markdown report generation failed for {record_id}: {e.message}")
        raise HTTPException(
            status_code=500,
            detail={
                "error": "generation_failed",
                "message": f"生成 Markdown 报告失败: {e.message}"
            }
        )
    except Exception as e:
        logger.error(f"获取 Markdown 报告失败: {e}", exc_info=True)
        raise HTTPException(
            status_code=500,
            detail={
                "error": "internal_error",
                "message": f"获取 Markdown 报告失败: {str(e)}"
            }
        )

    if markdown_content is None:
        raise HTTPException(
            status_code=404,
            detail={
                "error": "not_found",
                "message": f"未找到 id/query_id={record_id} 的分析记录"
            }
        )

    return MarkdownReportResponse(content=markdown_content)